+++ test/debug.jl
@@ -1,4 +1,4 @@
-
+@info "running debug.jl tests"
 @testset "test ac opf pbs" begin
 
     @testset "5-bus coupled meshed infeasible case" begin