+++ test/mld.jl
@@ -1,3 +1,4 @@
+@info "running mld.jl tests"
 @testset "test mld" begin
     @testset "test ACP" begin
         @testset "5-bus case" begin