+++ test/opf_bf.jl
@@ -1,3 +1,4 @@
+@info "running opf_bf.jl tests"
 @testset "test linearised distflow opf_bf" begin
     @testset "3-bus case" begin
         mp_data = PowerModels.parse_file("$(pms_path)/test/data/matpower/case3.m")