+++ CHANGELOG.md
@@ -2,6 +2,7 @@ PowerModelsDistribution.jl Change Log
 ===================================
 
 ### staged
+- Fix bug in OpenDSS parser on Transformers (#162)
 - FIX: OpenDSS parsing of `switch=y` property on lines (#161)
 
 ### v0.6.0